PSCApp/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "django>=6.0",
    "django-environ>=0.12.0",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.29.0",
    "django-cors-headers>=4.9.0",
    "whitenoise>=6.11.0",
    # "mysqlclient",
    "psycopg2-binary",
    "gunicorn>=23.0.0",
    "pillow>=12.1.0",
    "channels>=4.3.2",
    "django-redis>=6.0.0",
    "celery>=5.6.1",
    "django-filter>=25.2",
    "dj-rest-auth[with-social]>=5.0.2",
    "django-allauth>=0.57.2",
    "faker>=40.1.0",
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["."]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]
